Move the filter logic to the separate file
