+++ js/modules/filter.js
@@ -1,0 +1,29 @@
+const filterForm = document.querySelector('.map__filters');
+const filterFormFieldsets = filterForm.querySelectorAll('fieldset');
+const filterFormSelects = filterForm.querySelectorAll('.map__filter');
+
+
+/* States */
+const deactivateFilter = () => {
+  filterForm.classList.add('ad-form--disabled');
+
+  for(const fieldset of filterFormFieldsets) {
+    fieldset.setAttribute('disabled', '');
+  }
+  for(const select of filterFormSelects) {
+    select.setAttribute('disabled', '');
+  }
+};
+
+const activateFilter = () => {
+  filterForm.classList.remove('ad-form--disabled');
+
+  for(const fieldset of filterFormFieldsets) {
+    fieldset.removeAttribute('disabled');
+  }
+  for(const select of filterFormSelects) {
+    select.removeAttribute('disabled', '');
+  }
+};
+
+export {activateFilter, deactivateFilter};
