+++ js/modules/map.js
@@ -1,5 +1,6 @@
 import {generateAdvertisingMarkup} from './generate-advertising-markup.js';
-import {activateForm, activateFilter} from './form.js';
+import {activateForm} from './form.js';
+import {activateFilter} from './filter.js';
 import {mapOptions} from './map-options.js';
 
 const COUNT_SHOWED_ADVS = 10;
